代码修改

+++ mp-api/src/main/java/com/kulongtai/mpstore/entity/Card.java
@@ -37,7 +37,7 @@ package com.kulongtai.mpstore.entity;
 
             @ApiModelProperty(value = "卡券消费码")
         @TableField("card_no")
-    private Integer cardNo;
+    private Long cardNo;
 
             @ApiModelProperty(value = "用户id")
         @TableField("user_id")
